requirements.txt
# --- Ядро бота ---
aiogram==3.10.0            # Фреймворк Telegram-бота (укажите свежую 3.x)
redis==5.0.7               # Клиент Redis (для FSM и подписок)

# --- Сбор данных и Аналитика ---
ccxt==4.3.40               # Асинхронный API бирж (Binance)
pandas-ta==0.4.71b0        # Библиотека для Технического Анализа (RSI, MACD) - версия 0.3.14b недоступна в PyPI
pandas>=2.2.2              # Обработка данных (необходим для pandas-ta)
numpy>=2.2.6               # Вычисления (необходим для pandas-ta 0.4.71b0)

# --- Сеть и Вспомогательные ---
aiohttp==3.9.5             # Асинхронный HTTP-клиент (для DeepSeek и ccxt)
apscheduler==3.10.4        # Асинхронный планировщик задач

# --- Конфигурация ---
python-dotenv==1.0.1       # Загрузка .env файлов
